+++ tests/integration/main.rs
@@ -1,3 +1,4 @@
+mod error;
 mod report;
 pub(crate) mod util;
 mod wrap;